Rename module file
