+++ scripts/validation/fixes/assets.js
